+++ packages/core/types/index.d.ts
@@ -1,4 +1,4 @@
-export { AddonRegister, AddonGet, Addon, Addons, ResponseHandler, RemoveResponseHandler, SideEffect, Cache, RawParams, Params, RouteProps, ResponseSetters, FinishProps, EveryMatchFn, InitialMatchFn, FinishMatchFn } from './interface';
+export { AddonRegister, AddonGet, Addon, Addons, ResponseHandler, RemoveResponseHandler, SideEffect, Cache, RawParams, Params, RouteProps, ResponseSetters, ResponseBuilder, EveryMatchFn, InitialMatchFn, ResponseMatchFn } from './interface';
 export { Route, RouteDescriptor, Title, ParamParser, ParamParsers } from './route';
 export { Response } from './response';
 export { CuriConfig, ConfigOptions } from './curi';